+++ src/engine/index.ts
@@ -13,3 +13,4 @@
  */
 export * from './time';
 export * from './market';
+export * from './economy';